Update CHANGELOG.md
Definitely a nit, but we don't want to advertise write-catalog as a feature.

Long term it'll only be for users who need to access the legacy artifact, which we would also want to discourage. Short term, it'll be run under the hood and there are no recommend user facing impacts. We primarily introduced this so we could replace docs-generate. Broadly, the goal is to hydrate everything needed for your metadata automatically with build and run with the new platform rather than rely on extra commands for users to see their metadata hydrated in the platform.

https://dbt-labs.slack.com/archives/C094BCVA264/p1758564681222229