Standardize relation configs and changesets (#7018)
* Created relation config v2

* Add custom diffs and tests to config_v2

* Add `RelationComment` component

* Add `ColumnComments` component

* Add `Constraints` component

* Add `LiquidClustering`

This was already full of TODOs with no tests, I just ported it over to
the new format.

* Add `PartitionBy` component

* Add `Query` component

This was also full of TODOs and I simply ported.

* Add `Refresh` config component

* Add `RelationTags` components

* Add `TblProperties` component

* Fix import warnings

* Add `RelationConfigLoader` that works with `dyn ComponentConfigLoader`

Adds a `dyn ComponentConfigLoader` that is basically just a type-erased
passthrough to all the existing `from_remote_state()` and
`from_remote_state()` methods.

I added a `RelationConfigLoader` that can dynamically load a relation
type and all its components based on its inner type-erased loaders.

* Reuse mock tables in tests

* Reuse dbt model mocks across component tests

* Add `ColumnTags` component

* Move away from `TypeId` into custom `type_name`

Some components might share the same type, like a
`SimpleComponentConfigImpl<T>`, and then their type ID would be the same
even though the component should be a different one.

Now each component carries it's own `type_name` so we can have two
components of the same type meaning different things.

* Add type-erased `new()` and self-less `type_name()`

* Invert order of `RelationConfig` diff

Didn't realize it was (prev, next) instead of (next, prev) as in all
other places.

* Add view relation type test infra for relation types

* Test results between v1 and v2

This commit adds a way to perform an `Eq` comparison between v1 and v2
changesets and components.

This will ensure that our migration will behave the same as the v1
implementation, and that all keys are the same etc.

When we're done with this, we can delete this code and the integration
tests for each relation type will remain :)

* Add materialized view relation type and tests

* Add incremental table relation type and tests

* Add streaming table relation type and tests

* Make `requires_full_refresh` more flexible

Calculating a big AND between all components was not flexible enough.
Turns out Databricks might full refresh also depending on the
materialization type and on the component type and other present
components.

I removed the `requires_full_refresh` from the components and instead
made each `RelationConfig` hold a custom `requires_full_refresh()`
function that will get invoked with the map of components.

Then I added a global `requires_full_refresh()` in `mod.rs` to make it
super clear what things require/don't require full refresh.

* Move to new folder

* Add allowlist of v1 errors

* Fix clippy

* Fix streaming table full refresh requirements

* Add python reference comments

* Felipe's suggestions

* Rename `diff()` to `diff_from()`

GitOrigin-RevId: 4f8366e12f5318547fb9fddde27cde287180a8c0